Rename cds_to_mp3audiobook to cds_to_audiobook,
and instead of using lame, use ffmpeg, which makes it easier to switch
between formats as well as their options.

It looks like it handles joint stereo/mono a little differently, but
the files are the same size and they seem to sound the same?

The goal was that aac files should sound better, and hopefully I'll be
able to get big files in the iphone.