audiobook_tool: improvements to getting the length of a file
If the file isn't a wav file, convert it to a wav file and base the length off the wav file.

Using ffmpeg/ffplay to do this proves to be very accurate.

Use ffplay to for play_chapter, since that was the only requirement we had for mplayer,
obsoleting it's use.

Don't specify the number of seconds to play for play_chapter since ffplay doesn't support it,
and it wasn't very useful anyways.

Do more argument verification for commands that get passed a load of audio files

Do more error checking when getting the length of a file, and use exceptions to report those errors